+++ src/components/footer/footer.tsx
@@ -1,9 +1,7 @@
 import classes from "./footer.module.scss";
 import sponsors from "@/constants/sponsors";
 import SponsorLogos from "@/components/footer/sponsorLogos";
-
-const yearNow = new Date().getFullYear();
-const copyrights = `${yearNow} Games Store | All rights reserved | Terms Of Service | Privacy`;
+import { copyrights } from "@/constants/other";
 
 const Footer = () => (
   <footer className={classes.footer}>