+++ src/components/header/header.tsx
@@ -1,11 +1,17 @@
-import Navbar from "./navbar";
+import { FC } from "react";
 import classes from "./header.module.scss";
+import routes from "@/constants/routes";
+import otherImages from "@/constants/otherImages";
+import { storeName } from "@/constants/other";
+import Navbar from "@/components/header/navbar";
 
-const storeName = "Games $tore";
-
-const Header = () => (
+const Header: FC = () => (
   <header className={classes.header}>
-    <Navbar title={storeName} />
+    <nav className={classes.navbar}>
+      <img className={classes.store__logo} src={otherImages.storeLogo.path} alt={otherImages.storeLogo.description} />
+      <h1 className={classes.store__title}>{storeName}</h1>
+      <Navbar routes={routes} />
+    </nav>
   </header>
 );