+++ src/components/layout/layout.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import Header from "@/components/header/header";
+import Header from "../header/header";
 import Footer from "@/components/footer/footer";
 
 const Layout: FC = ({ children }) => (