+++ src/components/pages/products/products.tsx
@@ -1,9 +1,17 @@
 import Searchbar from "@/ui/searchbar/searchbar";
+import classes from "@/components/pages/home/home.module.scss";
+import products from "../../../../server/data/products";
+import GamesCard from "@/ui/gamesCard/gamesCard";
+import Container from "@/ui/container/container";
 
 const Products = () => (
   <div className="products__page">
-    <h1>Products page</h1>
     <Searchbar />
+    <Container id={classes.newGames} title="Games" isCard>
+      {products.map((product) => (
+        <GamesCard product={product} key={product.id} />
+      ))}
+    </Container>
   </div>
 );