+++ src/ui/button/button.tsx
@@ -3,7 +3,7 @@ import classes from "@/ui/button/button.module.scss";
 
 const Button: FC = ({ children }) => (
   <div className={classes.button__container}>
-    <button type="button" className={classes.button}>
+    <button type="submit" className={classes.button}>
       {children}
     </button>
   </div>